loan_approval_app/requirements.txt
streamlit
tensorflow
google-cloud-aiplatform